refactor: split identifiers into simple and advanced
The reason for this is to optimize the parser so it doesn't need to do a bunch of look ahead when
trying to parse things such as return statements. Another example is when trying to determine if a
case expression is a component identifier or the end of the case.