feat: pass type related tokens to nodes instead of converting types
This felt dirty to begin with and I wasn't too happy with converting a string to an actual `Type`
in the middle of building the AST. I've decided I will verify valid types in the semantic checking
stage of the compiler instead of invalid types being a syntax error. This also means I will not
need to feed the lexer a list of types that are valid to be defined in specific ways.