feat: basic implementation of a semantic analyzer
This implementation is still very much a WIP and currently only verifies if a scripts signature is
valid according to its trigger type. Things that need to be reported are all handled through the
`Diagnostics` class and is meant to be used throughout the whole compilation where each step checks
if there were any errors before continuing.