feat: create a WrappedType interface for more complex types
This type will be used for labels, procs, clientscripts, params, game vars, and any other type that
can have a different type depending on how it's referenced. For example, when setting up the
transmit list of a hook, you specify the varp without the `%` symbol. This kind of reference should
resolve to what essentially would be `varp<type>`, where `type` the return type if accessed using a
`%`.