feat: allow `null` to be assignable to long
This also removes the MetaType.NULL type since that was a weird way to approach
the `null` literal to begin with. We now rely on the type hint being defined to
know what the return type of `null` should be. If there is no type hint, then
`null` defaults to a return type of `int`.