feat: verify and store subjects that reference something
This adds proper checks how triggers should handle the subject passed to them. It is possible to
have a trigger that is just a simple name, a global subject, or a subject that refers to a category
or some other type. Category subjects must be prefixed with `_` and anything prefixed with that
also attempts to verify as a category.