feat: rework JoinedStringExpression to use StringPart
This is mostly prep work to support special tags, such as `<p,name>` and `<sprite=name>`. The consequence of this change is that the bytecode for strings with only tags now is split at every tag. In the future there will hopefully be a way to enable string optimizations that can optimize strings if wanted.