feat: optimizations
Removes the unnecessary caching of instructions and splits up pointer
calculation by type and req/set/corrupt. Splitting pointer calculations
means that less time is spent on pointers that are never even needing to
be checked.

This is mostly a WIP still, I will override this commit in the future :)